Begin migration of nacho.yml to windspigot.yml
