+++ WindSpigot-Server/src/main/java/dev/cobblesword/nachospigot/protocol/MinecraftPipeline.java
@@ -1,5 +1,6 @@
 package dev.cobblesword.nachospigot.protocol;
 
+import ga.windpvp.windspigot.config.WindSpigotConfig;
 import io.netty.buffer.ByteBufAllocator;
 import io.netty.channel.ChannelConfig;
 import io.netty.channel.ChannelInitializer;
@@ -28,9 +29,9 @@ public class MinecraftPipeline extends ChannelInitializer<SocketChannel> {
 	protected void initChannel(SocketChannel channel) {
 		try {
 			ChannelConfig config = channel.config();
-			config.setOption(ChannelOption.TCP_NODELAY, NachoConfig.enableTCPNODELAY);
-			config.setOption(ChannelOption.TCP_FASTOPEN, NachoConfig.modeTcpFastOpen);
-			config.setOption(ChannelOption.TCP_FASTOPEN_CONNECT, NachoConfig.enableTcpFastOpen);
+			config.setOption(ChannelOption.TCP_NODELAY, WindSpigotConfig.enableTCPNODELAY);
+			config.setOption(ChannelOption.TCP_FASTOPEN, WindSpigotConfig.modeTcpFastOpen);
+			config.setOption(ChannelOption.TCP_FASTOPEN_CONNECT,WindSpigotConfig.enableTcpFastOpen);
 			config.setOption(ChannelOption.IP_TOS, 0x18); // [Nacho-0027] :: Optimize networking
 			config.setAllocator(ByteBufAllocator.DEFAULT);
 		} catch (Exception ignored) {
