+++ WindSpigot-Server/src/main/java/net/minecraft/server/BlockLeaves.java
@@ -4,6 +4,7 @@ import java.util.Random;
 
 import org.bukkit.event.block.LeavesDecayEvent; // CraftBukkit
 
+import ga.windpvp.windspigot.config.WindSpigotConfig;
 import me.elier.nachospigot.config.NachoConfig;
 
 public abstract class BlockLeaves extends BlockTransparent {
@@ -144,7 +145,7 @@ public abstract class BlockLeaves extends BlockTransparent {
 	}
 
 	private void e(World world, BlockPosition blockposition) {
-		if (!NachoConfig.leavesDecayEvent) {
+		if (!WindSpigotConfig.leavesDecayEvent) {
 			return;
 		}
