+++ WindSpigot-Server/src/main/java/net/minecraft/server/EntityPotion.java
@@ -9,6 +9,7 @@ import org.bukkit.craftbukkit.entity.CraftLivingEntity;
 import org.bukkit.entity.LivingEntity;
 // CraftBukkit end
 
+import ga.windpvp.windspigot.config.WindSpigotConfig;
 import me.elier.nachospigot.config.NachoConfig;
 
 public class EntityPotion extends EntityProjectile {
@@ -28,7 +29,7 @@ public class EntityPotion extends EntityProjectile {
 		super(world, entityliving);
 		this.item = itemstack;
 		// IonSpigot start - Lag Compensated Potions
-		if (entityliving instanceof EntityPlayer && NachoConfig.lagCompensatedPotions) {
+		if (entityliving instanceof EntityPlayer && WindSpigotConfig.lagCompensatedPotions) {
 			((EntityPlayer) entityliving).potions.add(this);
 			compensated = true;
 		}
