+++ WindSpigot-Server/src/main/java/net/minecraft/server/PacketPlayInUseEntity.java
@@ -2,6 +2,7 @@ package net.minecraft.server;
 
 import java.io.IOException;
 
+import ga.windpvp.windspigot.config.WindSpigotConfig;
 import me.elier.nachospigot.config.NachoConfig;
 
 public class PacketPlayInUseEntity implements Packet<PacketListenerPlayIn> {
@@ -60,6 +61,6 @@ public class PacketPlayInUseEntity implements Packet<PacketListenerPlayIn> {
 
 	@Override
 	public boolean instant() {
-		return NachoConfig.instantPlayInUseEntity;
+		return WindSpigotConfig.instantPlayInUseEntity;
 	}
 }
