+++ WindSpigot-Server/src/main/java/net/minecraft/server/PlayerList.java
@@ -184,7 +184,7 @@ public abstract class PlayerList {
 		this.l = new IpBanList(PlayerList.b);
 		this.operators = new OpList(PlayerList.c);
 		// [Nacho-0037] Add toggle for "Faster Operator"
-		if (NachoConfig.useFastOperators) {
+		if (WindSpigotConfig.useFastOperators) {
 			for (OpListEntry value : this.operators.getValues()) {
 				this.fastOperator.add(value.getKey().getId());
 			}
@@ -1212,7 +1212,7 @@ public abstract class PlayerList {
 	public void addOp(GameProfile gameprofile) {
 		this.operators.add(new OpListEntry(gameprofile, this.server.p(), this.operators.b(gameprofile)));
 		// [Nacho-0037] Add toggle for "Faster Operator"
-		if (NachoConfig.useFastOperators) {
+		if (WindSpigotConfig.useFastOperators) {
 			this.fastOperator.add(gameprofile.getId());
 		}
 		// CraftBukkit start
@@ -1226,7 +1226,7 @@ public abstract class PlayerList {
 	public void removeOp(GameProfile gameprofile) {
 		this.operators.remove(gameprofile);
 		// [Nacho-0037] Add toggle for "Faster Operator"
-		if (NachoConfig.useFastOperators) {
+		if (WindSpigotConfig.useFastOperators) {
 			this.fastOperator.remove(gameprofile.getId());
 		}
 
@@ -1240,13 +1240,13 @@ public abstract class PlayerList {
 
 	public boolean isWhitelisted(GameProfile gameprofile) {
 		// [Nacho-0037] Add toggle for "Faster Operator"
-		return !this.hasWhitelist || (NachoConfig.useFastOperators ? this.fastOperator.contains(gameprofile.getId())
+		return !this.hasWhitelist || (WindSpigotConfig.useFastOperators ? this.fastOperator.contains(gameprofile.getId())
 				: this.operators.d(gameprofile)) || this.whitelist.d(gameprofile);
 	}
 
 	public boolean isOp(GameProfile gameprofile) {
 		// [Nacho-0037] Add toggle for "Faster Operator"
-		return (NachoConfig.useFastOperators ? this.fastOperator.contains(gameprofile.getId())
+		return (WindSpigotConfig.useFastOperators ? this.fastOperator.contains(gameprofile.getId())
 				: this.operators.d(gameprofile))
 				|| this.server.T() && this.server.worlds.get(0).getWorldData().v()
 						&& this.server.S().equalsIgnoreCase(gameprofile.getName())
