+++ WindSpigot-Server/src/main/java/net/minecraft/server/World.java
@@ -2345,9 +2345,9 @@ public abstract class World implements IBlockAccess {
 		double d0 = 1.0D / ((axisalignedbb.d - axisalignedbb.a) * 2.0D + 1.0D);
 		double d1 = 1.0D / ((axisalignedbb.e - axisalignedbb.b) * 2.0D + 1.0D);
 		double d2 = 1.0D / ((axisalignedbb.f - axisalignedbb.c) * 2.0D + 1.0D);
-		double d3 = (1.0D - ((NachoConfig.enableFastMath ? FastMath.floor(1.0D / d0) : Math.floor(1.0D / d0)) * d0))
+		double d3 = (1.0D - ((WindSpigotConfig.enableFastMath ? FastMath.floor(1.0D / d0) : Math.floor(1.0D / d0)) * d0))
 				/ 2.0D;
-		double d4 = (1.0D - ((NachoConfig.enableFastMath ? FastMath.floor(1.0D / d2) : Math.floor(1.0D / d2)) * d2))
+		double d4 = (1.0D - ((WindSpigotConfig.enableFastMath ? FastMath.floor(1.0D / d2) : Math.floor(1.0D / d2)) * d2))
 				/ 2.0D;
 
 		if (d0 >= 0.0D && d1 >= 0.0D && d2 >= 0.0D) {
